[Concurrency] Added a fan handler to retrieve information easily
